split out callback and styling functionality into own classes
