+++ browsergui/elements/_hascallbacks.py
@@ -1,0 +1,39 @@
+import collections
+
+class NoSuchCallbackError(Exception):
+  """Raised when trying to remove a nonexistent callback from an Element."""
+  pass
+
+class HasCallbacks(object):
+  def __init__(self):
+    super(HasCallbacks, self).__init__()
+    self.callbacks = collections.defaultdict(list)
+
+  def add_callback(self, event_type, callback):
+    """Arranges for ``callback`` to be called whenever the Element handles an event of ``event_type``.
+
+    :type event_type: str
+    :type callback: a function of one argument (the event being handled)
+    """
+    self.callbacks[event_type].append(callback)
+    if self.gui is not None:
+      self.gui.note_callback_added(self, event_type, callback)
+
+  def remove_callback(self, event_type, callback):
+    """Stops calling ``callback`` when events of ``event_type`` are handled.
+
+    For parameter information, see :func:`add_callback`.
+    """
+    if callback not in self.callbacks[event_type]:
+      raise NoSuchCallbackError(event_type, callback)
+    self.callbacks[event_type].remove(callback)
+    if self.gui is not None:
+      self.gui.note_callback_added(self, event_type, callback)
+
+  def handle_event(self, event):
+    """Calls all the callbacks registered for the given event's type.
+
+    :type event: dict
+    """
+    for callback in self.callbacks[event['type']]:
+      callback(event)
